+++ frontend/public/src/entry/header.js
@@ -39,11 +39,11 @@ const renderHeader = () => {
   )
 }
 
-const renderEnrollSection = (courseId, element, reduxStore) => {
+const renderEnrollSection = (courseId, programId, element, reduxStore) => {
   ReactDOM.render(
     <AppContainer>
       <Provider store={reduxStore}>
-        <ProductDetailEnrollApp courseId={courseId} />
+        <ProductDetailEnrollApp courseId={courseId} programId={programId} />
       </Provider>
     </AppContainer>,
     element
@@ -67,10 +67,12 @@ document.addEventListener("DOMContentLoaded", function() {
   const enrollSectionEl = document.getElementById("productDetailEnrollment")
   const upsellCardEl = document.getElementById("upsellCard")
   const courseIdEl = document.getElementById("courseId")
-  if (enrollSectionEl && courseIdEl && upsellCardEl) {
+  const programIdEl = document.getElementById("programId")
+  if (enrollSectionEl && (programIdEl || courseIdEl) && upsellCardEl) {
     const productDetailStore = configureStore()
-    const courseId = courseIdEl.value
-    renderEnrollSection(courseId, enrollSectionEl, productDetailStore)
+    const courseId = courseIdEl ? courseIdEl.value : undefined
+    const programId = programIdEl ? programIdEl.value : undefined
+    renderEnrollSection(courseId, programId, enrollSectionEl, productDetailStore)
     renderUpsellCard(courseId, upsellCardEl, productDetailStore)
   }
 })
